requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector-python==8.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.21.0
protobuf==4.25.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.29
statsmodels==0.14.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
torch==2.2.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
xgboost==2.0.3
zipp==3.18.1
